RF: Merge simplesetup.py with network.py.
