+++ pelita/player/team.py
@@ -14,7 +14,7 @@ import zmq
 from . import AbstractTeam
 from .. import libpelita, layout
 from ..exceptions import PlayerDisconnected, PlayerTimeout
-from ..simplesetup import ZMQConnection, ZMQClientError, ZMQReplyTimeout, ZMQUnreachablePeer, DEAD_CONNECTION_TIMEOUT
+from ..network import ZMQConnection, ZMQClientError, ZMQReplyTimeout, ZMQUnreachablePeer
 
 
 _logger = logging.getLogger(__name__)
